# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CollAPI - API de Cobranzas",
        "description": "API REST para el sistema de gestión de cobranzas",
        "contact": {
            "email": "soporte@collapi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/directions/{id}": {
            "get": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Obtener direccion",
                "description": "Obtiene una direccion por su ID",
                "operationId": "getDirection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la direccion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Direccion obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Direccion obtenida correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "integer"
                                                },
                                                "direction": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "province": {
                                                    "type": "string"
                                                },
                                                "canton": {
                                                    "type": "string"
                                                },
                                                "parish": {
                                                    "type": "string"
                                                },
                                                "neighborhood": {
                                                    "type": "string"
                                                },
                                                "latitude": {
                                                    "type": "string"
                                                },
                                                "longitude": {
                                                    "type": "string"
                                                },
                                                "client_name": {
                                                    "type": "string"
                                                },
                                                "client_ci": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Direccion no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Actualizar direccion",
                "description": "Actualiza una direccion existente",
                "operationId": "updateDirection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la direccion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "direction": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Av. Principal 456"
                                    },
                                    "type": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "TRABAJO"
                                    },
                                    "province": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Guayas"
                                    },
                                    "canton": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Guayaquil"
                                    },
                                    "parish": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Tarqui"
                                    },
                                    "neighborhood": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Kennedy"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "-2.1894"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "-79.8891"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Direccion actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Direccion actualizada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Direccion no encontrada"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Eliminar direccion",
                "description": "Elimina una direccion",
                "operationId": "deleteDirection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la direccion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Direccion eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Direccion eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Direccion no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}